requirements.txt
django==3.1.4
django-ckeditor==6.0.0
django-environ==0.4.5
gunicorn==20.0.4
psycopg2==2.8.6
whitenoise==5.2.0
